Extract shared K8s utilities into k8s_utils module
